requirements.txt
asgiref==3.3.1
boto3==1.16.52
botocore==1.19.52
coverage==5.4
cryptography==3.3.1
dj-database-url==0.5.0
Django==3.1.5
django-allauth==0.44.0
django-crispy-forms==1.10.0
django-libs==2.0.3
django-localflavor==3.0.1
django-phone-field==1.8.1
django-storages==1.11.1
factory-boy==3.2.0
Faker==6.3.0
gunicorn==20.0.4
jmespath==0.10.0
mixer==7.1.2
oauthlib==3.1.0
Pillow==8.1.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
PyJWT==2.0.0
python-stdnum==1.14
python3-openid==3.2.0
pytz==2020.5
requests-oauthlib==1.3.0
s3transfer==0.3.4
sqlparse==0.4.1
stripe==2.55.1
text-unidecode==1.3
